Updated Jump Controls and Completed Level 1 Platforsm
Created various different platforms for players to jump onto to complete the first level. Also added more coins throughout for players to retrieve as they complete the level. Had issues with jumping mechanisms on the platforms I created but resolved this in the scripts by editing the OnCollisionEnter method to allow jumping mechanism to work on platform with tags of both "Floor" and "Platforms".