Added UI hints, to guide user, finalised coins/treasure, added scripts to end level
When user reaches 10 coins (all of the treasure in the first level, after reaching the final large piece of treasure), they are prompted that they have progressed. The user receives reminders to collect ALL the coins. Tutorial "blocks" have been added which, on collision, tell the user hints e.g. turn off gravity. The game checks if the user is in a starting position using 3d vectors, coordinates, as well as if they are in the first scene to display welcome message.