Implemented lives counter, ensured conditions are met to enter next level
Made it such that if a player falls off, lives counter is decremented by one (until it reaches 0 - where a game over screen will be displayed in future). The portal will now either tell a player to collect more coins or - if this condition is met - allow them to enter the next level.