Updated level 1 difficulty and implemented improved gravity controls
I have updated the speed of the animated blocks so that it is easier to pass through, based off of the playtesting workshop were some testers commented on its difficulty, and so completed animation speed for level 1 moving objects is 4 seconds (previous was 2). Also updated the gravity controls such that there is a 5 second period in which the low gravity mode is implemented. This creates a sense of difficulty which is adaptable based on the platforms players are jumping on to.