Level 2 traps, and sprint/crouch controls
updated level 2 to implement different types of traps in which users will need to jump over these traps/ crouch underneath the traps. also implemented an animated trap npc that will cause the user to lose a life if they make contact. reduced the default speed from 20f to 15f. also reduced speed multipler from x2 to x1.5. Also implemented a speed multiplier of x0.5 for when the player is crouched. debugged the code so that the player cannot sprint while crouched.