Level 3 and enemy NPCs
Created a new scene for level 3 called thirdLevel. at the moment it is just a plan that contains the Player, and a cube called Enemy. The Enemy Cube is an NPC which is currently being developed. So far the NPC can track the movements of the player using public variables including sight, and detection range and targets any object that has the tag "Player", in this case the user. Currently has a default speed of 2 and sight range of 20 for testing purposes. Please note, Level 3 scene is not fully implemented yet as doing testing for the NPC

Co-Authored-By: umairYousaf803 <146939881+umairYousaf803@users.noreply.github.com>
Co-Authored-By: ismail275 <146841997+ismail275@users.noreply.github.com>
Co-Authored-By: Danyaal Zar <zardanyaal@gmail.com>