Completed basic enemy NPC models
Completed the model for NPC in which when thet arrack, they do 20 damage onto the player. the play starts off with 100 health, and each time they are attacked they lose 20hp. When the player hits 0hp, the LivesText script is called and decrements the lives by 1, the same way we lose lives if we fall off the map. Also allowed for respawn after losing a life from the NPC. NOTE: The TestingNPC Scene is a temporary scene I have created to do testing and will be edited as the 3rd level later on.

Co-Authored-By: umairYousaf803 <146939881+umairYousaf803@users.noreply.github.com>
Co-Authored-By: ismail275 <146841997+ismail275@users.noreply.github.com>
Co-Authored-By: Danyaal Zar <zardanyaal@gmail.com>