NPC Health
Created a model for the health points (located in EnemyScipts.cs) that stay attached above the enemy NPCs which are set at 100% as a default. Also fixed a bug where when the player is out of sight from the enemy npc, the npc would tilt over and essentially fall. We have resolved this issue by editing their rigib body constraints in which their rotations are frozen once we are out of range using .FreezeRotation("X/Y"). Also edited the colour of the health and the NPC itself to distinguish that it is an enemy NPC.

Co-Authored-By: umairYousaf803 <146939881+umairYousaf803@users.noreply.github.com>
Co-Authored-By: ismail275 <146841997+ismail275@users.noreply.github.com>
Co-Authored-By: Danyaal Zar <zardanyaal@gmail.com>