Friendly NPCs and debugging
Added multiple friendly NPCs in level 1 which act as guide for players on how to progress throughout the levels. Have begun adding on level 2, but did not finish. Also few debugging issues regarding multiples texts coming up relevant to the player on top of the screen rather than just the one set.