Settings menu (input sensitivity, volume), redid logic for lives collectibles, fixed bugs, implemented basic inventory
Settings menu now included (accessible through main menu), which contains settings for input sensitivity (works) and game volume (which has not been implemented).

I fixed and redid the logic for the lives collectibles as these did not work. They do now.

Implemented a basic inventory for guns - user presses E to open and close this, and this will be complemented by the shop soon.