NavMesh Enemy NPC
Updated the Enemy NPC to use AI Pathfinding (Navmesh) which allows the enemy to find a valid route to the user based off of the area they are in. e.g., if there is a wall between the player and the enemy, the enemy will go around the wall rather than get stuck trying to walk through it. This has been implemented in the existing EnemyScript & there is an additional script called NavigationScript. Also adjusted the scripts so that they work together, were previous logic wasnt basing the movement off of NavMesh Agents & Surface, which now utilises this. Removed any conflict between previous and existing script.