Player Attack and NPC Health/Death
Edited PlayerController Script to include an attack feature which does a default damage of 10% when no weapons are equipped and its just his fist which works within a range of 5f. Also edited the EnemyScript such that their health corresponds to the attack damage taken from user input. Need to resolve a bug where the player doesnt have to directly aim at the npc to inflict damage