Jump & Health Points was not fixed
Tested the game to make sure jump and health mechanics were working but were not. Turns out Danyaals code for the sound broke the jumping sound effects and so had to make slight adjustments to fix the jumping issue. As for the Health issue, the Player originally lost their lives if they came into contact with an object tag of "EnemyNPC". Removed this tag and now health and lives are working properly.