Player can now only use guns in their inventory
If they press 1, 2, 3 and have bought the respective gun it will show (for level 3), they can shoot (need to implement shoot for gun 2 and 3 and add a tutorial prompt for how to use them). The shop scene now leads to level 3 - to test the flow described start at level 1, collect all coins and enter the shop and buy some guns.