Navmesh Surface and Agent
Finally corrected all the navmesh issues. Main issue was the npcs rigid body had to be changed to enable kinematics so that the AI would use the navmesh agent and its script in order to travel up slopes which are shown on the present level. previously couldnt travel up the slope. Also fixed a bug where the platforms on level 3 didnt allow users to jump. (NOTE: TESTINGNPC LEVEL has been renamed to level 3 and this level is to get an idea of how to combat against enemies)