Jump Fix & Improvements
Where we have the isGrounded variable in PlayerController.cs, we had to edit the OnCollisionEnter method and remove the collision pattern between the player and the floor/platform. Instead, we are keeping this method just for contact death with the traps. In addition there is a new method OnCollisionStay, which checks if the player remains grounded to the platform/floor. Previously we had an error where the game would recognise the player as not grounded after standing on a slope even though they were clearly standing on the platform. with this new method, we have added a float, slopeangle which states that if the slope is below 45 deg., the player is still grounded. on top of fixing this, we also resolved an issue where the player could essentially jump if they were hanging onto a wall.