Fixed Error where players can hang/stick onto walls
There has been a very big error in which players can essentially stick to walls due to the friction force between the players capsule collider and game object. to rectify this issue, I have created a new physics material with low frictional forces. In doing so, it creates no friction when a player walks into an object and falls towards the walls of an object, and so instead of hanging/sticking to the wall, they fall instead.