Navigation, PCG & Hitbox
Edited the NavigationScript so that the patrol points area is smaller as the enemies spawned in ridiculous locations. Also edited the PCG so we reduced the total no. enemies that spawn in based on difficulty. Finally edited the PlayerControllerScript in which we edited the PerformAttack method so that we use Raycasting. this is important as it allows the user to directly attack the enemies rather than cause "area damage" which was a serious bug in the game, and so the centre dot must be aiming at the enemies to inflict damage. Had an issue with the hitbox of the enemies, so edited to colliders to make the hitbox match with the enemy transform.