LEVEL 4 COMPLETION
further developed level 4 as the final level, few adjustments required - moving platform animations need to be added, creating a script called moving platform that can be added as a component to allow a user to stand on them, must ensure the onTrigger is selected. Weapons need to be made available for use by the player and replace the final large coin with a ruby to signify treasure found and game won at the end. Enemies included with collectibles, add navmesh space