Final Level
Added Navmesh to final level. Corrected the enemy NPC such that we use kinematics for their navmesh agent. this fixed a bug where the agents would move with irregular speeds. Also created a MovingPlatformController script which should allow the player to move with the animated platforms in sync rather than falling off. This temporarily deactivates the material for the player in which their frictional force is low. Still need to attach this to moving objects but will complete afterwards.